+++ api-keys-auth-sample/package.json
@@ -5,8 +5,7 @@
   "license": "ISC",
   "repository": "git@github.com/Livestream/livestream-api-samples",
   "devDependencies": {
-    "crypto": "^0.0.3",
-    "express": "^4.14.0",
-    "moment": "^2.14.1"
+    "request": "^2.74.0",
+    "express": "^4.14.0"
   }
 }
