+++ api-keys-auth-sample/server.js
@@ -1,9 +1,8 @@
 var express = require('express');
-var crypto = require('crypto');
-var moment = require('moment');
+var request = require('request');
 
 const PORT=8080;
-const API_SECRET = "[YOUR_SECRET_API_KEY]";
+const API_SECRET = "[YOUR_API_KEY]";
 
 var app = express();
 
@@ -18,29 +17,28 @@ app.use(function (req, res, next) {
 
 app.use(express.static('client_application'));
 
-function generateToken(scope, timestamp) {
-    var hmac = crypto.createHmac("md5", API_SECRET);
-    
-    var tokenString = API_SECRET + ":" + scope + ":" + timestamp
-    hmac.update(tokenString);
-    var crypted = hmac.digest("hex");
 
-    return crypted;
-};
 
-app.get("/token", function(req, res) {
-    var currentTime = moment().format('x');
+app.get("/accounts", function(req, res) {
 
-    var SCOPE = 'readonly';
-    
-    var token = generateToken(SCOPE, currentTime);
+    var options = {
+        url: 'https://livestreamapis.com/v1/accounts',
+        auth: {
+            username: API_SECRET
+        }
+    }
 
-    console.log('scope:' + SCOPE);
-    console.log('token:' + token);
-    console.log('timestamp: ' + currentTime);
+    request.get(options, function(err, response, body) {
+        if (err) {
+            console.error('livestream api call failed ', err)
+            return res.send(err)
+        }   
 
-    res.setHeader('Content-Type', 'application/json');
-    res.send({'timestamp': currentTime, 'token': token});
+        console.log('data received ' + body);
+        res.setHeader('Content-Type', 'application/json');
+        res.send(body)
+    })
+
 });    
 
 app.listen(PORT, function () {
