+++ secure-token-auth-sample/README.md
@@ -1,14 +1,17 @@
-### Example of Client Server making Livestream Rest API calls for your Angular App.
+### Example of Angular Application making Livestream Rest API calls
+
+![Flow](flow.png)
 
 ### Prerequisite
 - Install `node`
 
 ### Run your server & client app on same port
 
-1. `cd secure-token-auth-sample/`
+1. `cd api-keys-auth-sample/`
 2. edit `[YOUR_API_KEY]` with your secret key in `server.js`
-2. run `npm install`
-3. start server `node server.js`
-4. Open `http://localhost:8080/index.html` and you should see your account data. 
+3. edit `[YOUR_CLIENT_ID]` with your client id in `client_application/app.js`
+3. run `npm install`
+4. start server `node server.js`
+5. Open `http://localhost:8080/index.html` and you should see your account data. 
 
 For more information please visit the documentation - https://livestreamapis.com/docs
