+++ secure-token-auth-sample/client_application/app.js
@@ -1,5 +1,7 @@
 angular.module('sampleApp', [])
-    .directive('accounts', ['DataService', function(dataService){
+    .directive('accounts', ['LsService', 'TokenService', function(lsService, tokenService){
+
+        var CLIENT_ID = [YOUR_CLIENT_ID];
 
         return {
             restrict: 'E', 
@@ -8,12 +10,24 @@ angular.module('sampleApp', [])
                 $scope.accounts = {}
 
                 //Call to your server
-                dataService.getAccounts({
+                tokenService.getToken({
                     success: function(data, status, headers, config) {
                         console.log(status)
-                        $scope.accounts = data;
+
+                        //Call to livestream api server
+                        lsService.getAccounts(data.token, CLIENT_ID, data.timestamp, {
+                            success: function(data, status, headers, config) {
+                                console.log(status)
+
+                                $scope.accounts = data;
+
+                            },
+                            error: function(data, status, headers, config) {
+                                console.log(status)
+                            }
+                        })
+
                     },
-                    
                     error: function (data, status, headers, config) {
                         console.log(status)
                     }
